mv LineSplitting to src/utils.jl
