+++ src/GeoMakie.jl
@@ -38,7 +38,6 @@ include("geojson.jl") # GeoJSON/GeoInterface support
 include("conversions.jl")
 include("data.jl")
 include("utils.jl")
-include("LineSplitting.jl")
 
 @reexport using Colors, Makie
 export Proj
